Export a variable to set GCP service account
By default a new service account is created to run nodes for the GKE
cluster, and this account gets the IAM role logging.logWriter assigned to
operate correctly.

To simplify the role management when the module is used from a GH
Action, the variable create_service_account is exposed, in order to
force the use of the same service account configured in the Action.

Doing so, it is sufficient to configure that single Service Account with
the correct IAM Role.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>